+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - Esborrar un Rol d'un Usuari amb `removerole < @user > < @rol >`. Més info amb la comanda `help removerole`.
 - Afegida nova comanda `setcounter` que adjudica el canal de text com a contador de membres.
     - El comptador s'actualitza cada 12h o manualment amb la comanda `refresh` que només poden executar els Administradors.
+- Nova comanda `morsify` / `demorsify` que tradueix un text a Morse i al revés.
 
 \[v1.2\]